requirements.txt
openai>=1.35.0
mistralai>=1.2.0
python-dotenv>=1.0.1
Pillow>=10.0.0
gradio>=4.0.0
